Refactored views out. see #21.
