+++ paperwork/urls.py
@@ -11,12 +11,12 @@ app_name = 'paperwork'
 urlpatterns = [
     url(r'^$', views.home, name='home'),
     url(r'^client_info_types/$', views.client_info_types),
-    url(r'^clients/$', views.clients),
-    url(r'^deliverables/$', views.deliverables),
+    url(r'^clients/$', views.view_clients),
+    url(r'^deliverables/$', views.view_deliverables),
     url(r'^new_deliverable/$', views.new_deliverable),
     url(r'^deliverable/(?P<deliverable_id>[0-9]+)/$',
         views.view_deliverable, name="deliverable"),
-    url(r'^tasks/$', views.tasks),
+    url(r'^tasks/$', views.view_tasks),
     url(r'^dpc/$', views.dpc),
     url(r'^login/$', views.login_page),
     url(r'^make_tasks/$', views.make_tasks)
